requirements.txt
flask==1.0.2
flask-restful==0.3.6
redis==2.10.6
python-decouple==3.1
schematics==2.1.0
aiohttp==3.4.4
pyjwt==1.7.0
sqlalchemy==1.2.12